Split SvgRadialGradientServer
